+++ android/src/main/java/com/appsflyer/appsflyersdk/AppsflyerSdkPlugin.java
@@ -8,6 +8,8 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.os.Looper;
 import android.util.Log;
+import java.util.concurrent.atomic.AtomicBoolean;
+import androidx.annotation.NonNull;
 
 import com.appsflyer.AFLogger;
 import com.appsflyer.AppsFlyerConsent;
@@ -22,6 +24,7 @@ import com.appsflyer.share.LinkGenerator;
 import com.appsflyer.share.ShareInviteHelper;
 import com.appsflyer.internal.platform_extension.Plugin;
 import com.appsflyer.internal.platform_extension.PluginInfo;
+import com.appsflyer.attribution.AppsFlyerRequestListener;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -80,7 +83,7 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
     private Boolean isFacebookDeferredApplinksEnabled = false;
     private Boolean isSetDisableAdvertisingIdentifiersEnable = false;
     private Map<String, Map<String, Object>> mCallbacks = new HashMap<>();
-    
+
     PluginRegistry.NewIntentListener onNewIntentListener = new PluginRegistry.NewIntentListener() {
         @Override
         public boolean onNewIntent(Intent intent) {
@@ -200,7 +203,7 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
 
     @Override
     public void onMethodCall(MethodCall call, Result result) {
-        if(activity == null){
+        if (activity == null) {
             Log.d("AppsFlyer", "Activity isn't attached to the flutter engine");
             return;
         }
@@ -369,11 +372,52 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
         result.success(null); // indicate that the method invocation is complete
     }
 
-    private void startSDK(MethodCall call, Result result) {
-        AppsFlyerLib instance = AppsFlyerLib.getInstance();
-        instance.start(activity);
-    }
+    /**
+     * Initiates the AppsFlyer SDK. The AtomicBoolean isResultSubmitted ensures the result is
+     * only submitted once, preventing the "Reply already submitted" exception in Flutter.
+     */
+    private void startSDK(MethodCall call, final Result result) {
+        final AtomicBoolean isResultSubmitted = new AtomicBoolean(false);
 
+        try {
+            final AppsFlyerLib instance = AppsFlyerLib.getInstance();
+            String afDevKey = call.argument(AppsFlyerConstants.AF_DEV_KEY);
+            instance.start(activity, afDevKey, new AppsFlyerRequestListener() {
+                @Override
+                public void onSuccess() {
+                    new Handler(Looper.getMainLooper()).post(new Runnable() {
+                        @Override
+                        public void run() {
+                            if (!isResultSubmitted.getAndSet(true)) {
+                                mMethodChannel.invokeMethod("onSuccess", null);
+                                result.success(null);
+                            }
+                        }
+                    });
+                }
+
+                @Override
+                public void onError(final int errorCode, final String errorMessage) {
+                    new Handler(Looper.getMainLooper()).post(new Runnable() {
+                        @Override
+                        public void run() {
+                            if (!isResultSubmitted.getAndSet(true)) {
+                                HashMap<String, Object> errorDetails = new HashMap<>();
+                                errorDetails.put("errorCode", errorCode);
+                                errorDetails.put("errorMessage", errorMessage);
+                                mMethodChannel.invokeMethod("onError", errorDetails);
+                                result.error(String.valueOf(errorCode), errorMessage, null);
+                            }
+                        }
+                    });
+                }
+            });
+        } catch (Exception e) {
+            if (!isResultSubmitted.getAndSet(true)) {
+                result.error("UNEXPECTED_ERROR", e.getMessage(), null);
+            }
+        }
+    }
     public void setConsentData(MethodCall call, Result result) {
         Map<String, Object> arguments = (Map<String, Object>) call.arguments;
         Map<String, Object> consentDict = (Map<String, Object>) arguments.get("consentData");
@@ -394,6 +438,7 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
 
         result.success(null);
     }
+
     private void enableTCFDataCollection(MethodCall call, Result result) {
         boolean shouldCollect = (boolean) call.argument("shouldCollect");
         AppsFlyerLib.getInstance().enableTCFDataCollection(shouldCollect);
@@ -401,13 +446,13 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
     }
 
     private void addPushNotificationDeepLinkPath(MethodCall call, Result result) {
-        if(call.arguments != null){
+        if (call.arguments != null) {
             ArrayList<String> depplinkPath = (ArrayList<String>) call.arguments;
             String[] depplinkPathArr = depplinkPath.toArray(new String[depplinkPath.size()]);
             AppsFlyerLib.getInstance().addPushNotificationDeepLinkPath(depplinkPathArr);
         }
         result.success(null);
-    }    
+    }
 
     private void setDisableNetworkData(MethodCall call, Result result) {
         boolean disableNetworkData = (boolean) call.arguments;
@@ -420,10 +465,10 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
     }
 
     private void setOutOfStore(MethodCall call, Result result) {
-            String sourceName = (String) call.arguments;
-            if (sourceName != null) {
-                AppsFlyerLib.getInstance().setOutOfStore(sourceName);
-            }
+        String sourceName = (String) call.arguments;
+        if (sourceName != null) {
+            AppsFlyerLib.getInstance().setOutOfStore(sourceName);
+        }
         result.success(null);
     }
 
@@ -438,14 +483,14 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
     private void setPartnerData(MethodCall call, Result result) {
         String partnerId = (String) call.argument("partnerId");
         HashMap<String, Object> partnerData = (HashMap<String, Object>) call.argument("partnersData");
-        if(partnerData != null){
+        if (partnerData != null) {
             AppsFlyerLib.getInstance().setPartnerData(partnerId, partnerData);
         }
         result.success(null);
     }
 
     private void setSharingFilterForPartners(MethodCall call, Result result) {
-        if(call.arguments != null){
+        if (call.arguments != null) {
             ArrayList<String> partnersInput = (ArrayList<String>) call.arguments;
             String[] partners = partnersInput.toArray(new String[partnersInput.size()]);
             AppsFlyerLib.getInstance().setSharingFilterForPartners(partners);
@@ -465,7 +510,7 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
 
     private void enableFacebookDeferredApplinks(MethodCall call, Result result) {
         isFacebookDeferredApplinksEnabled = (boolean) call.argument("isFacebookDeferredApplinksEnabled");
-        
+
         if (isFacebookDeferredApplinksEnabled) {
             AppsFlyerLib.getInstance().enableFacebookDeferredApplinks(true);
         } else {
@@ -484,7 +529,7 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
         String errorMsg = null;
         Bundle bundle;
 
-        if(pushPayload == null){
+        if (pushPayload == null) {
             Log.d("AppsFlyer", "Push payload is null");
             return;
         }
@@ -509,13 +554,14 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
             errorMsg = "The activity is null. Push payload has not been sent!";
         }
 
-        if(errorMsg != null){
+        if (errorMsg != null) {
             Log.d("AppsFlyer", errorMsg);
             return;
         }
 
         result.success(null);
     }
+
     private static Bundle jsonToBundle(JSONObject jsonObject) throws JSONException {
         Bundle bundle = new Bundle();
         Iterator iter = jsonObject.keys();
@@ -526,6 +572,7 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
         }
         return bundle;
     }
+
     private void setOneLinkCustomDomain(MethodCall call, Result result) {
         ArrayList<String> brandDomains = (ArrayList<String>) call.arguments;
         String[] brandDomainsArray = brandDomains.toArray(new String[brandDomains.size()]);
@@ -768,11 +815,11 @@ public class AppsflyerSdkPlugin implements MethodCallHandler, FlutterPlugin, Act
     private void setUserEmails(MethodCall call, Result result) {
         List<String> emails = call.argument("emails");
         int cryptTypeInt = call.argument("cryptType");
-        
+
         AppsFlyerProperties.EmailsCryptType cryptType = null;
-        if (cryptTypeInt == 0){
+        if (cryptTypeInt == 0) {
             cryptType = AppsFlyerProperties.EmailsCryptType.NONE;
-        } else if (cryptTypeInt == 1){
+        } else if (cryptTypeInt == 1) {
             cryptType = AppsFlyerProperties.EmailsCryptType.SHA256;
         } else {
             throw new InvalidParameterException("You can use only NONE or SHA256 for EmailsCryptType on android");