+++ lib/src/callbacks.dart
@@ -10,6 +10,8 @@ const _channel = MethodChannel('callbacks');
 typedef MultiUseCallback = void Function(dynamic msg);
 typedef UDLCallback = void Function(DeepLinkResult deepLinkResult);
 typedef CancelListening = void Function();
+typedef RequestSuccessListener = void Function();
+typedef RequestErrorListener = void Function(int errorCode, String errorMessage);
 
 Map<String, MultiUseCallback> _callbacksById = <String, void Function(dynamic)>{
 };
@@ -21,6 +23,7 @@ Future<void> _methodCallHandler(MethodCall call) async {
       try {
         dynamic callMap = jsonDecode(call.arguments);
         switch (callMap["id"]) {
+          case "startSDK":
           case "onAppOpenAttribution":
           case "onInstallConversionData":
           case "validatePurchase":